LayoutTests/networkinformation/basic-all-types-of-events-expected.txt

Tests the basic operation of all NetworkInfo events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


webkitnetworkinfochange event is raised
PASS connection.bandwidth is bandwidth
PASS connection.metered is metered
PASS successfullyParsed is true

TEST COMPLETE
